requirements.txt
# requirements.txt records the full set of dependencies for development
accelerate
codetiming
datasets
dill
flash-attn
hydra-core
liger-kernel
numpy
pandas
datasets
peft
pyarrow>=15.0.0
pybind11
pylatexenc
pre-commit
ray[default]
tensordict<=0.6.2
torchdata
transformers
vllm==0.6.3.post1
wandb
absl-py==2.1.0
accelerate==1.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-cors==0.7.0
aiosignal==1.3.1
airportsdata==20241001
alabaster==1.0.0
alibabacloud_credentials==0.3.6
alibabacloud-tea==0.4.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
anytree==2.12.1
archspec==0.2.3
astor==0.8.1
asttokens==3.0.0
attrs==24.2.0
av==14.4.0
babel==2.16.0
bitsandbytes==0.45.0
blake3==1.0.1
blinker==1.9.0
boltons==24.0.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
codetiming==1.4.0
colorama==0.4.6
coloredlogs==15.0.1
colorful==0.5.6
comm==0.2.2
common-io==0.4.1+tunnel
compressed-tensors==0.9.2
conda==24.9.2
conda-libmamba-solver==24.9.0
conda-package-handling==2.4.0
conda_package_streaming==0.11.0
cupy-cuda12x==13.4.1
Cython==3.0.11
dataclasses-json==0.6.7
datasets==3.2.0
debugpy==1.8.11
decorator==5.1.1
deepspeed==0.15.0
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
docutils==0.21.2
einops==0.8.0
email_validator==2.2.0
evaluate==0.4.3
executing==2.1.0
faiss-cpu==1.9.0
fastapi==0.115.6
fastapi-cli==0.0.7
fastrlock==0.8.3
fbgemm_gpu==1.0.0+cu121
filelock==3.16.1
flash-attn==2.7.0.post2
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.2.0
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.0
google-auth==2.37.0
googleapis-common-protos==1.66.0
graphlearn==1.3.0
grpcio==1.67.1
grpcio-tools==1.62.3
h11==0.14.0
h2==4.1.0
hjson==3.1.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.2
humanfriendly==10.0
hydra-core==1.3.2
hyperframe==6.0.1
identify==2.6.1
idna==3.10
imagesize==1.4.1
importlib_metadata==8.5.0
iniconfig==2.0.0
interegular==0.3.3
intervaltree==3.1.0
iopath==0.1.9
ipykernel==6.29.5
ipython==8.31.0
isort==5.13.2
jedi==0.19.2
Jinja2==3.1.3
jiter==0.8.2
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lark==1.2.2
latex2sympy2_extended==1.10.1
libcst==1.5.0
libmambapy==1.5.10
liger_kernel==0.5.2
lightning-utilities==0.11.8
linkify-it-py==2.0.3
llguidance==0.7.20
llvmlite==0.43.0
lm-format-enforcer==0.10.11
loralib==0.1.2
mamba==1.5.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
math-verify==0.7.0
mathruler==0.1.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.15.0
menuinst==2.1.2
mistral_common==1.5.4
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
myst-parser==4.0.0
narwhals==1.21.1
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.3
nodeenv==1.9.1
numba==0.60.0
numpy==1.26.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.4.0
omegaconf==2.3.0
openai==1.79.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.10.0.84
openrlhf==0.5.5.post2
optimum==1.23.3
orjson==3.10.15
outlines==0.1.11
outlines_core==0.1.26
packaging==24.1
pandas==2.2.3
parameterized==0.9.0
parso==0.8.4
partial-json-parser==0.2.1.1.post4
peft==0.14.0
pexpect==4.9.0
pillow==10.4.0
pip==24.2
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
pre_commit==4.0.1
prometheus_client==0.21.1
prometheus-fastapi-instrumentator==7.0.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.0
pyairports==2.1.1
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycosat==0.6.6
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pyfg==0.3.4
Pygments==2.18.0
pylatexenc==2.10
pynvml==12.0.0
pyodps==0.12.0
pyre-check==0.9.21
pyre-extensions==0.0.31
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qwen-vl-utils==0.0.11
ray==2.46.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.14.6
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.5.0
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==8.1.3
sphinx-rtd-theme==3.0.1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
starlette==0.41.3
streamlit==1.41.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensordict==0.6.2
tensorrt-cu12==10.3.0
tensorrt-cu12-bindings==10.3.0
tensorrt-cu12-libs==10.3.0
termcolor==2.5.0
TestSlide==2.7.1
textual==1.0.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
torch_tensorrt==2.5.0
torchaudio==2.6.0
torchdata==0.11.0
torchmetrics==1.0.3
torchrec==1.0.0+cu121
torchvision==0.21.0
tornado==6.4.2
tqdm==4.66.5
traitlets==5.14.3
transformers==4.51.3
transformers-stream-generator==0.0.5
triton==3.1.0
truststore==0.9.2
typeguard==2.13.3
typer==0.15.4
typing_extensions==4.12.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2024.2
uc-micro-py==1.0.3
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
verl==0.2.0.dev0
virtualenv==20.27.1
vllm==0.8.2
wandb==0.19.1
watchdog==6.0.0
watchfiles==1.0.3
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.1.1
wheel==0.44.0
wrapt==1.17.0
xformers==0.0.29.post2
xgrammar==0.1.16
xxhash==3.5.0
yarl==1.17.1
zipp==3.21.0
zstandard==0.23.0
